Fix scrolling for keyboard navigation
In some cases `scrollIntoView` was behaving erratic, i.e. the scroll
position of the whole page changed although the results were already
visible fully.

We therefore change the scrolling to alter `scrollTop` directly. The
scrolling is "lazy" so that it only changes when a selected item either
vanishes at the top or the bottom. Changing the selection within the
visible items in the list does not change the scroll position.